Fixed some problems that may appear in future
