+++ main.cpp
@@ -4,6 +4,33 @@
 #include "gameobject.h"
 #include "drawing.h"
 
+void DrawingExploreState(GameObject& player){
+    if (IsKeyDown(KEY_W)){
+        player.SetMoveUp();
+        player.Move();
+    }
+    if (IsKeyDown(KEY_S)){
+        player.SetMoveDown();
+        player.Move();
+    }
+    if (IsKeyDown(KEY_U)){
+        player.radius += 2;
+    }
+    if (IsKeyDown(KEY_D)){
+        player.radius -= 2;
+    }
+    
+   
+
+    BeginDrawing();
+    ClearBackground(BLACK);
+
+    player.Update();
+
+    DrawFPS(10, 10);
+    EndDrawing();
+}
+
 int main() {
     const int width = 900;
     const int height = 900;
